+++ changelog.md
@@ -1,3 +1,10 @@
+### Version 2.1.5
+
+-   Changed behavior from throwing error in case `qrbox.width` or `qrbox` is larger
+    than the width of the root element. In such cases the dimension will automatically
+    be truncated to the size of root element and will throw a warning based on verbosity
+    settings. This should address [issue#357](https://github.com/mebjas/html5-qrcode/issues/357)
+
 ### Version 2.1.4
 
 #### Huge thanks to [Ben Richardson](https://github.com/ben-gy) for one time sponsorship!!